Append keyswap hook on `isearch-mode-end-hook'
For hooks where one is an entry point and the other is an exit, I think
it's good behaviour to make what you have either wrap the entire current
action or just the original function.
e.g. when adding a setup and teardown hook into `isearch-mode-hook' and
`isearch-mode-end-hook' respectively, one should be at the start of the
hook and the other at the end of its hook.